Style : 피그마 참고하여 수정
